+++ services/automation/CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ### Features
 
-- **services:** give every service its own version and changelog (c2e4e2c)
+- **services:** give every service its own version and changelog (#45) (5bc91b8)
 
 ## v0.3.2 (2026-09-15)